Change all time references to qint32 instead of 64, add basic handling for position delivery; DOES NOT COMPILE

+++ software/GUI/WAG_bands/recordingcontroller.h
@@ -25,7 +25,7 @@ private:
 signals:
 
 public slots:
-    void addSnapshotToMotion(qint64, PositionSnapshot);
+    void addSnapshotToMotion(qint32, PositionSnapshot);
     void catchVoiceControlCommand(MessageType);
     void toggleVoiceControl(bool voiceControlOn);
